Move HoverService into platform
Fixes #206228
