+++ src/vs/editor/standalone/browser/standaloneServices.ts
@@ -9,7 +9,7 @@ import '../../../platform/undoRedo/common/undoRedoService.js';
 import '../../common/services/languageFeatureDebounce.js';
 import '../../common/services/semanticTokensStylingService.js';
 import '../../common/services/languageFeaturesService.js';
-import '../../browser/services/hoverService/hoverService.js';
+import '../../../platform/hover/browser/hoverService.js';
 import '../../browser/services/inlineCompletionsService.js';
 
 import * as strings from '../../../base/common/strings.js';
