+++ src/vs/platform/hover/browser/hover.css
