+++ src/vs/workbench/workbench.common.main.ts
@@ -116,7 +116,7 @@ import './services/authentication/browser/authenticationMcpService.js';
 import './services/authentication/browser/dynamicAuthenticationProviderStorageService.js';
 import './services/authentication/browser/authenticationQueryService.js';
 import './services/accounts/common/defaultAccount.js';
-import '../editor/browser/services/hoverService/hoverService.js';
+import '../platform/hover/browser/hoverService.js';
 import './services/assignment/common/assignmentService.js';
 import './services/outline/browser/outlineService.js';
 import './services/languageDetection/browser/languageDetectionWorkerServiceImpl.js';
